+++ src/output.c
@@ -16,7 +16,7 @@ struct Output_Data output_data;
 
 void output(char *buf, enum eOutputSubsystem subsystem, enum eOutputLevel level) {
    int i, l=strlen(buf);
-   for(i=l-1;i>=0 && !isprint(buf[i]); l--, i--);
+   for(i=l-1;i>=0 && !isprint((int)buf[i]); l--, i--);
    if((output_data.mask[eOutputChannelUART][subsystem] & (1<<level)) != 0)
       UART0_Transmit(buf, l, 1);
